refact: reorganized macros/ast-term-derive
